+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.*;
